# project/storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Custom Bikes",
        "contact": {
            "name": "Developer Contact",
            "url": "https://www.example.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/api/navires": {
            "get": {
                "tags": [
                    "Navires"
                ],
                "summary": "List Navires",
                "responses": {
                    "200": {
                        "description": "Success: List all Navires",
                        "schema": {
                            "$ref": "#/definitions/Navire"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Navires"
                ],
                "summary": "Create Navire",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Json format",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Navire"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Success: A Newly Created Navire",
                        "schema": {
                            "$ref": "#/definitions/Navire"
                        }
                    },
                    "422": {
                        "description": "Missing mandatory field"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "405": {
                        "description": "Invalid HTTP Method"
                    }
                }
            }
        },
        "/api/navires/{id}": {
            "get": {
                "tags": [
                    "Navires"
                ],
                "summary": "Get Navire by Id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Display the specified Navire by id.",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success: Return the Navire",
                        "schema": {
                            "$ref": "#/definitions/Navire"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "405": {
                        "description": "Invalid HTTP Method"
                    }
                }
            },
            "put": {
                "tags": [
                    "Navires"
                ],
                "summary": "Update Navire",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Update the specified Navire by id.",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Json format",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Navire"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success: Return the Navire updated",
                        "schema": {
                            "$ref": "#/definitions/Navire"
                        }
                    },
                    "422": {
                        "description": "Missing mandatory field"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "405": {
                        "description": "Invalid HTTP Method"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Navires"
                ],
                "summary": "Delete Navire",
                "description": "Delete the specified Navire by id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Navire id to delete",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Not found"
                    },
                    "405": {
                        "description": "Invalid HTTP Method"
                    },
                    "204": {
                        "description": "Success: successful deleted"
                    }
                }
            }
        },
        "/api/profils": {
            "get": {
                "tags": [
                    "Profils"
                ],
                "summary": "List Profils",
                "responses": {
                    "200": {
                        "description": "Success: List all Profils",
                        "schema": {
                            "$ref": "#/definitions/Profil"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Profils"
                ],
                "summary": "Create Profil",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Json format",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Profil"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Success: A Newly Created Profil",
                        "schema": {
                            "$ref": "#/definitions/Profil"
                        }
                    },
                    "422": {
                        "description": "Missing mandatory field"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "405": {
                        "description": "Invalid HTTP Method"
                    }
                }
            }
        },
        "/api/profils/{id}": {
            "get": {
                "tags": [
                    "Profils"
                ],
                "summary": "Get Profil by Id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Display the specified Profil by id.",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success: Return the Profil",
                        "schema": {
                            "$ref": "#/definitions/Profil"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "405": {
                        "description": "Invalid HTTP Method"
                    }
                }
            },
            "put": {
                "tags": [
                    "Profils"
                ],
                "summary": "Update Profil",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Update the specified Profil by id.",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Json format",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Profil"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success: Return the Profil updated",
                        "schema": {
                            "$ref": "#/definitions/Profil"
                        }
                    },
                    "422": {
                        "description": "Missing mandatory field"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "405": {
                        "description": "Invalid HTTP Method"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Profils"
                ],
                "summary": "Delete Profil",
                "description": "Delete the specified Profil by id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Profil id to delete",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Not found"
                    },
                    "405": {
                        "description": "Invalid HTTP Method"
                    },
                    "204": {
                        "description": "Success: successful deleted"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Navire": {
            "required": [
                "libelle",
                "contenanceTotale"
            ],
            "properties": {
                "libelle": {
                    "description": "libelle d\u00e9crivant le Navire",
                    "type": "string",
                    "example": "Malt"
                },
                "contenanceTotale": {
                    "description": "libelle d\u00e9crivant le Navire",
                    "type": "string",
                    "example": "Malt"
                }
            }
        },
        "Profil": {
            "required": [
                "libelle"
            ],
            "properties": {
                "libelle": {
                    "description": "libelle d\u00e9crivant le profil",
                    "type": "string",
                    "example": "Root"
                }
            }
        },
...
    }
}